refactor(fixtures): collapse scalar fixture wrappers behind scalar_fixture!
Recovered from orphaned commit 0a60f71 (dropped by a reset during the
stacked-PR shuffle). The eql_v2_int4 fixture file was ~95% boilerplate
shared with future scalar types: the spec() builder, the fixture-gen
generator test, and the property-test module differed only in name, the
Rust plaintext type, and the value const.

Add a scalar_fixture!(name, ty, values) macro that stamps out all three.
MIN/MAX in the signed-extremes test derive from <$ty>. The int2 hunk
from the original commit is dropped — int2 is not on this branch.

Test-infra only; no caller-observable change. Fixture property tests
pass (3/3); generate() still compiles under --features fixture-gen.
